cloud_functions/copy-msrb-replay-to-gcs/requirements.txt
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
Deprecated==1.2.15
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-cloud-secret-manager==2.22.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.68.1
grpcio-status==1.68.1
idna==3.10
importlib_metadata==8.5.0
numpy==1.26.4
opentelemetry-api==1.29.0
packaging==24.2
pandas==1.5.3
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.17.0
urllib3==2.3.0
wrapt==1.17.0
zipp==3.21.0

# google-cloud-bigquery
# google-cloud-bigquery-storage
# google-cloud-storage
# google-cloud-secret-manager
# google-cloud-logging
# pytz
# numpy==1.26.4
# pandas==1.5.3
